pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "drone-vps"
version = "0.1.0"
description = "Drone Visual Positioning System — GPS-denied navigation via satellite image matching"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
    "opencv-python-headless>=4.8",
    "pydantic>=2.0",
    "pyserial>=3.5",
]

[project.optional-dependencies]
onboard = [
    "onnxruntime>=1.17",
    "faiss-cpu>=1.7",
]
programmer = [
    "torch>=2.0",
    "kornia>=0.7",
    "faiss-cpu>=1.7",
    "aiohttp>=3.9",
    "click>=8.0",
    "pygeotile>=1.0",
    "tqdm",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
vps-onboard = "onboard.main:main"
vps-program = "programmer.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
